backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-dotenv
openai>=1.3.0
arcadepy
httpx
scikit-learn
python-dateutil
lxml
lxml_html_clean

rapidfuzz
beautifulsoup4
trafilatura
